+++ sfftkrw/__init__.py
@@ -4,6 +4,6 @@ import os
 BASE_DIR = os.path.dirname(__file__)
 VALID_EXTENSIONS = ['sff', 'xml', 'hff', 'h5', 'hdf5', 'json']
 
-SFFTKRW_VERSION = 'v0.5.2.dev0'
+SFFTKRW_VERSION = 'v0.5.2.dev1'
 
 SFFTKRW_ENTRY_POINT = 'sff-rw'